Improve API documentation with search and filtering
Added interactive filtering and searching to the OpenAPI specification documentation within the Confluence plugin, along with versioning features.  Changes include updates to `CommentEntity.java`, `OpenApiMacro.java`, `OpenApiParser.java`, `OpenApiRenderer.java`, `CommentResource.java`, `CommentServiceImpl.java`, `atlassian-plugin.xml`, `openapi-documentation.js`, and several new files related to versioning and search functionality (`ApiSpecVersionEntity.java`, `ApiSpecVersion.java`, `ApiSpecVersionResource.java`, `ApiSpecVersionServiceImpl.java`, `OpenApiDiffUtil.java`, `openapi-search.css`, `openapi-search.js`, `openapi-versions.css`, `openapi-versions.js`, `pluginIcon.svg`, `pluginLogo.svg`).  The `@Scanned` annotation was replaced with `@Component` in several classes.  The `parentId` type in `CommentEntity` was changed from `Long` to `Integer`.

+++ src/main/java/com/example/confluence/ao/CommentEntity.java
@@ -44,8 +44,8 @@ public interface CommentEntity extends Entity {
     String getAuthorAvatarUrl();
     void setAuthorAvatarUrl(String authorAvatarUrl);
     
-    Long getParentId(); // For replies, the ID of the parent comment
-    void setParentId(Long parentId);
+    Integer getParentId(); // For replies, the ID of the parent comment
+    void setParentId(Integer parentId);
     
     Integer getHighlightStartOffset(); // For annotation highlighting
     void setHighlightStartOffset(Integer highlightStartOffset);
